feat:리퀘스트 메시지 파싱작업을 타입별로 하도록 인터페이스 활용 get과 post메서드가 각각 메시지 파싱을 하도록 인터페이스 활용
